requirements.txt
cffi==1.15.1
cryptography==41.0.4
mysql-connector-python==8.0.33
protobuf==3.20.3
pycparser==2.21
